@@ -1,4 +1,4 @@
-import Vue from 'vue';
+import { createApp } from 'vue';
 import OdsMap from './map.vue';
 import fetchGeoJsonData from './geoJsonUtils';
 
@@ -16,13 +16,17 @@ document.addEventListener('DOMContentLoaded', () => {
   const mapElement11 = document.getElementById('ods-map11');
 
   if (mapElement) {
-    Vue.component('ods-map', OdsMap);
-    const vm1 = new Vue();
-    vm1.$mount(mapElement);
+    const app1 = createApp({
+      name: 'OdsMapApp1',
+      components: { OdsMap },
+    });
+    app1.mount(mapElement);
   }
 
   if (mapElement2) {
-    const vm2 = new Vue({
+    const app2 = createApp({
+      name: 'OdsMapApp2',
+      components: { OdsMap },
       data() {
         return {
           geoJsonData: null,
@@ -38,11 +42,13 @@ document.addEventListener('DOMContentLoaded', () => {
         }, 15000);
       },
     });
-    vm2.$mount(mapElement2);
+    app2.mount(mapElement2);
   }
 
   if (mapElement3) {
-    const vm3 = new Vue({
+    const app3 = createApp({
+      name: 'OdsMapApp3',
+      components: { OdsMap },
       data() {
         return {
           geoJsonData: null,
@@ -58,12 +64,13 @@ document.addEventListener('DOMContentLoaded', () => {
         }, 15000);
       },
     });
-    vm3.$mount(mapElement3);
+    app3.mount(mapElement3);
   }
 
   if (mapElement4) {
-    Vue.component('ods-map', OdsMap);
-    const vm4 = new Vue({
+    const app4 = createApp({
+      name: 'OdsMapApp4',
+      components: { OdsMap },
       mounted() {
         this.$refs.odsMap4.resize();
       },
@@ -72,11 +79,13 @@ document.addEventListener('DOMContentLoaded', () => {
         resize() {},
       },
     });
-    vm4.$mount(mapElement4);
+    app4.mount(mapElement4);
   }
 
   if (mapElement5) {
-    const vm5 = new Vue({
+    const app5 = createApp({
+      name: 'OdsMapApp5',
+      components: { OdsMap },
       data() {
         return {
           geoJsonData: null,
@@ -92,11 +101,13 @@ document.addEventListener('DOMContentLoaded', () => {
         }, 15000);
       },
     });
-    vm5.$mount(mapElement5);
+    app5.mount(mapElement5);
   }
 
   if (mapElement6) {
-    const vm6 = new Vue({
+    const app6 = createApp({
+      name: 'OdsMapApp6',
+      components: { OdsMap },
       data() {
         return {
           geoJsonData: null,
@@ -112,12 +123,13 @@ document.addEventListener('DOMContentLoaded', () => {
         }, 15000);
       },
     });
-    vm6.$mount(mapElement6);
+    app6.mount(mapElement6);
   }
 
   if (mapElement7) {
-    Vue.component('ods-map', OdsMap);
-    const vm7 = new Vue({
+    const app7 = createApp({
+      name: 'OdsMapApp7',
+      components: { OdsMap },
       data: () => ({
         points: [{ longitude: 10.74981, latitude: 59.913008, popupContent: '<h2>Bob the Builder</h2>' }],
       }),
@@ -128,11 +140,13 @@ document.addEventListener('DOMContentLoaded', () => {
         }, 10000);
       },
     });
-    vm7.$mount(mapElement7);
+    app7.mount(mapElement7);
   }
 
   if (mapElement8) {
-    const vm8 = new Vue({
+    const app8 = createApp({
+      name: 'OdsMapApp8',
+      components: { OdsMap },
       data() {
         return {
           geoJsonData: null,
@@ -155,12 +169,13 @@ document.addEventListener('DOMContentLoaded', () => {
         }, 15000);
       },
     });
-    vm8.$mount(mapElement8);
+    app8.mount(mapElement8);
   }
 
   if (mapElement9) {
-    Vue.component('ods-map', OdsMap);
-    const vm9 = new Vue({
+    const app9 = createApp({
+      name: 'OdsMapApp9',
+      components: { OdsMap },
       data: () => ({
         state: { longitude: 10.74981, latitude: 59.913008, zoom: 10, showPopups: true, autoFitToBounds: false },
       }),
@@ -171,18 +186,19 @@ document.addEventListener('DOMContentLoaded', () => {
         }, 15000);
       },
     });
-    vm9.$mount(mapElement9);
+    app9.mount(mapElement9);
   }
 
   if (mapElement10) {
-    Vue.component('ods-map', OdsMap);
-    const vm10 = new Vue({
+    const app10 = createApp({
+      name: 'OdsMapApp10',
+      components: { OdsMap },
       data: () => ({
         loadMap: false,
         state: { longitude: 10.74981, latitude: 59.913008, zoom: 10, showPopups: true, autoFitToBounds: false },
       }),
     });
-    vm10.$mount(mapElement10);
+    app10.mount(mapElement10);
   }
 
   if (mapElement11) {
@@ -230,4 +246,50 @@ document.addEventListener('DOMContentLoaded', () => {
     });
     vm11.$mount(mapElement11);
   }
+  if (mapElement11) {
+    const vm11 = new Vue({
+      data() {
+        return {
+          geoJsonData: null,
+          loadMap: false,
+        };
+      },
+      async created() {
+        try {
+          this.geoJsonData = await fetchGeoJsonData('https://ukeweb-public.s3.dualstack.eu-central-1.amazonaws.com/map/data/kindergarten-with-events-ods.geojson');
+        } catch (error) {
+          throw new Error('Error fetching GeoJSON data:', error);
+        }
+      },
+      methods: {
+        showMap() {
+          this.loadMap = true;
+        },
+        hideMap() {
+          this.loadMap = false;
+        },
+      },
+      async mounted() {
+        setTimeout(async () => {
+          try {
+            this.geoJsonData = await fetchGeoJsonData('https://ukeweb-public.s3.eu-central-1.amazonaws.com/map/data/featurecollection-with-popups-and-multi-variants.geojson');
+          } catch (error) {
+            throw new Error('Error fetching GeoJSON data:', error);
+          }
+        }, 15000);
+
+        const listViewBtn = document.getElementById('listViewBtn');
+        const mapViewBtn = document.getElementById('mapViewBtn');
+
+        listViewBtn.addEventListener('click', () => {
+          this.hideMap();
+        });
+        mapViewBtn.addEventListener('click', () => {
+          this.showMap();
+        });
+      },
+    });
+    vm11.$mount(mapElement11);
+  }
 });
+
